consolidate db creation into one script, and call it on heroku deploy
