--- db_mysql_setup.php
@@ -1,29 +1,0 @@
-<?php
-require_once('config/settings.php');
-
-$reports_sql = "CREATE TABLE IF NOT EXISTS `{$db_reports_table}` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `user_id` int(10) unsigned NOT NULL,
-  `course_id` int(10) unsigned NOT NULL,
-  `file_path` text NOT NULL,
-  `date_run` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  `errors` int(10) unsigned NOT NULL,
-  `suggestions` int(10) unsigned NOT NULL,
-  PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;";
-
-$users_sql = "CREATE TABLE IF NOT EXISTS `{$db_user_table}` (
-  `id` int(10) unsigned NOT NULL,
-  `api_key` varchar(255) NOT NULL,
-  `date_created` date NOT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `id` (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;";
-
-$dbh = include('lib/db.php');
-$sth = $dbh->query($reports_sql);
-$sth->execute();
-
-$sth = $dbh->query($users_sql);
-$sth->execute();
-
